pyproject.toml
[tool.poetry]
name = "minesweeper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygame-2048 = "^1.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
